Require that test targets have a host application
Without a host application, many of the generated accessibility properties in UIKit won't be populated correctly, which can lead to things silently failing in less-than-obvious ways. This adds an assertion at the beginning of the test methods to check that the host application exists, and adds a note to the README about this requirement.