Tile and stitch images when snapshotting large views
Views that are very large in at least one dimension will silently fail to render when using `drawHierarchy(in:afterScreenUpdates:)`, producing an empty snapshot of the view. This adds logic to draw tiled images of the view at a smaller size and stitch them together in order to produce the final snapshot.