Add a minimum macOS version that matches the SnapshotTesting dependency
The SnapshotTesting dependency specifies 10.15 as the minimum version. Even though this library doesn't support macOS, it's helpful to specify one so it can be consumed in mixed-platform packages. Today, if a mixed iOS/macOS package depends on AccessibilitySnapshot, SPM emits an error: "error: the library 'AccessibilitySnapshot' requires macos 10.13, but depends on the product 'SnapshotTesting' which requires macos 10.15; consider changing the library 'AccessibilitySnapshot' to require macos 10.15 or later, or the product 'SnapshotTesting' to require macos 10.13 or earlier."